chore($function): maintain car object and detach functions
Overall structural changes and separation of functions

Before:
-racingCar : package
 -Car : Car Object
 -RacingCar : do race and functions

After:
-racingCar : package
 -Car : Car Object
 -RacingHost: racing function detached
 -RacingCar: only do race start
