+++ README.md
@@ -17,21 +17,26 @@
 
 ## 객체
 
-- 자동차 객체
-  - 이름
-  - 위치 값
-  - 전진 함수
+### 자동차 객체(Car)
+  - [x] 이름(name)
+  - [x] 위치 값(position)
+  - [x] 전진 함수(forward)
     - 자동차의 위치 값을 증가 시키는 함수
-  - 현재 상태 출력 함수
+  - [x] 현재 상태 출력 함수(showPosition)
     - 현재 자동차의 상태 (위치값) 을 출력하는 함수
+  - [ ] 자동차의 현재 위치가 인수의 값보다 앞인지 판단하는 함수(isAheadOf)
+  - [ ] 자동차의 현재 위치가 인수의 값과 같은지 판단(isEqualOf)
 
-## 함수
+### 레이싱 관련 함수 제공 객체(RacingHost)
+  - [ ] 경주를 위한 자동차 객체 생성 및 자동차 저장 객체 리턴 함수(initCar)
+    - 사용자의 입력 값을 추출해 이름을 가져오고, 개수만큼 자동차를 생성하고, 자동차를 저장한 객체를 리턴.
+  - [x] 값 입력 함수(inputString)
+    - 사용자의 입력 값을 받아 저장하는 함수
+  - [ ] 경주 1회 수행 함수(racing)
+    - 경주 1회를 수행
+  - [x] 무작위 값 추첨 함수(draw)
+    - 0~9 사이 값을 추첨하는 함수
+  - [x] 우승자 계산 및 출력 함수(chooseWinner)
+    - 현재 자동차들의 위치값을 비교하여 최대값을 계산 후, 우승자를 결정 및 출력하는 함수
+  - [ ] 자동차 이름에 대한 예외처리 함수(carNameValidChk)
 
-- 값 입력 함수
-  - 사용자의 입력 값을 받아 저장하는 함수
-- 자동차 이름 추출 및 생성 함수
-  - 사용자의 입력 값을 추출해 이름을 가져오고, 개수만큼 자동차를 생성함
-- 무작위 값 추첨 함수
-  - 0~9 사이 값을 추첨하는 함수
-- 우승자 계산 및 출력 함수
-  - 현재 자동차들의 위치값을 비교하여 최대값을 계산 후, 우승자를 결정 및 출력하는 함수
