+++ src/test/java/racingCar/RacingCarTest.java
@@ -13,13 +13,13 @@ public class RacingCarTest {
     void racingTest() {
         System.setIn(new ByteArrayInputStream("pobi,woni,jun\n5".getBytes()));
         RacingCar racingCar = new RacingCar();
-        racingCar.race();
+        racingCar.start();
     }
 
     @Test
     void racingInputLengthErrorTest() {
         System.setIn(new ByteArrayInputStream("pobi,woni,november\n5".getBytes()));
         RacingCar racingCar = new RacingCar();
-        racingCar.race();
+        racingCar.start();
     }
 }
